+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,12 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.4] - Unreleased
+
+### Changed
+
+- Revised some error messages related to executing external utilities.
+
 ## [3.0.3] - 2025-10-18
 
 ### Changed
@@ -293,6 +299,7 @@ later.
 
 - Initial release of the shell
 
+[3.0.4]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.4
 [3.0.3]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.3
 [3.0.2]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.2
 [3.0.1]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.1